Casting index to int because apparently floats can't be strict compared to integers. Closes #2

+++ helper.php
@@ -219,7 +219,7 @@ class ModDwdwetterHelper
 	 */
 	public static function getDirection($grad)
 	{
-		$index = round($grad / 22.5);
+		$index = (int) round($grad / 22.5);
 		$index = ($index === 16) ? 0 : $index;
 
 		return Text::_('MOD_DWD_WETTERMODUL_DIRECTION_' . $index);
